+++ scripts/bench_vector_db_compare.sh
@@ -7,7 +7,7 @@ cd "$ROOT"
 RUN_DIR="${RUN_DIR:-/tmp/gomap_vector_db_compare_$(date +%Y%m%d_%H%M%S)}"
 PYTHON="${PYTHON:-python3}"
 VENV="${VENV:-$RUN_DIR/venv}"
-BACKENDS="${BACKENDS:-treedb,vectorlite,pgvector}"
+BACKENDS="${BACKENDS:-treedb,vectorlite}"
 DOCS="${DOCS:-10000}"
 DIMS="${DIMS:-64}"
 QUERIES="${QUERIES:-10000}"
